feat: auth module is now complete.
